refactor(TeacherProjectService): Move getBranchLetter to OneWorkgroupPerRowDataExportStrategy
