+++ src/assets/wise5/services/teacherProjectService.ts
@@ -870,21 +870,6 @@ export class TeacherProjectService extends ProjectService {
     return this.configService.getConfigParam('simulationProjectId') || -1;
   }
 
-  /**
-   * Get the branch letter in the node position string if the node is in a branch path
-   * @param nodeId the node id we want the branch letter for
-   * @return the branch letter in the node position if the node is in a branch path
-   */
-  getBranchLetter(nodeId) {
-    const nodePosition = this.getNodePositionById(nodeId);
-    const branchLetterRegex = /.*([A-Z])/;
-    const match = branchLetterRegex.exec(nodePosition);
-    if (match != null) {
-      return match[1];
-    }
-    return null;
-  }
-
   componentChanged(): void {
     this.componentChangedSource.next();
   }
